requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
pypdf
python-dotenv
IPython
ipykernel
llama-index-embeddings-gemini
streamlit 